requirements.txt
flask==3.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
tavily-python==0.5.0
markitdown==0.0.1a2
google-genai==1.5.1
google-generativeai==0.8.3
requests==2.32.4
google-adk==1.18.0
gunicorn==21.2.0
